Comment 관련 수정 및 error 수정

+++ src/main/resources/static/js/board/dataform/deleteBoard.js
@@ -1,5 +1,12 @@
 const deleteBoard = () => {
-    if(memberIdOfBoard === idOfLoginMember || typeOfLoginMember != "USER") {
+
+    findCommentList(boardId);
+    const commentList = document.getElementsByClassName("commentList")[0];
+    const commentCnt =commentList.getElementsByClassName("comment").length;
+    if(memberIdOfBoard === idOfLoginMember && commentCnt != 0) {
+        alert("게시글 삭제가 불가합니다. 관리자에게 문의해 주세요.");
+    }
+    if((memberIdOfBoard === idOfLoginMember && commentCnt == 0)|| typeOfLoginMember != "USER") {
         if(confirm("삭제하시겠습니까?")) {
             fetch("/api/board/delete/" + boardId, {
                 method : "DELETE"
@@ -7,7 +14,7 @@ const deleteBoard = () => {
                 if (response.ok) {
                     location.href = document.location.origin + "/complete/board?type=DELETE";
                 } else {
-                    throw new Error('오류가 발생하였습니다.');
+                    throw new Error(response.status + " 오류가 발생하였습니다.")
                 }
             }).catch(err => alert(err))
         }
